Sets a status in redis to check if user channel(s) is updating but needs bug fixes.

+++ feed_subscription/services.py
@@ -2,9 +2,12 @@ from typing import List, Optional
 
 import backoff
 from celery import shared_task, group
+from django.core.cache import cache
 from django.db import transaction
 
 from base_app.error_codes import ApplicationErrorException, ErrorCodes
+from feed_subscription.cache import is_user_channels_updating, is_user_channel_updating, set_user_channels_updating, \
+    set_user_channel_updating
 from feed_subscription.models import FeedChannel, Article
 from feed_subscription.selectors import try_get_channel_by_id, try_get_channel_by_rss_link, \
     get_existing_channel_by_id, get_user_existing_channels, get_channel_by_id
@@ -13,14 +16,14 @@ from users.selectors import get_user_by_id
 from utils.scraper import get_articles, get_channel_data
 
 
-def subscribe_to_channel(*args, user: ApplicationUser, rss_link: str) -> FeedChannel:
+def subscribe_to_channel(*args, user: ApplicationUser, rss_link: str, title: str) -> FeedChannel:
     with transaction.atomic():
         channel = try_get_channel_by_rss_link(user=user, rss_link=rss_link)
         if channel is not None and channel.deleted:
             channel.deleted = False
-            channel.save()
         elif channel is None:
-            channel = FeedChannel.objects.create(user=user, rss_link=rss_link)
+            channel = FeedChannel(user=user, rss_link=rss_link)
+        channel.title = title
         channel.save()
     update_all_user_channels.delay(user.id)
     return channel
@@ -47,32 +50,48 @@ def fetch_articles_task(rss_link: str, channel_id):
 
 @shared_task
 def update_all_user_channels(user_id: int):
-    user = ApplicationUser.objects.get(id=user_id)
-    channels = get_user_existing_channels(user=user)
-    g = group(fetch_articles_task.s(c.rss_link, c.id) for c in channels)
-    result = g().get()
-    Article.objects.filter(channel__in=channels).delete()
-    articles = []
-    for r in result:
-        if r is None:
-            continue
-        ar = _create_articles_from_task_result(result_list=r.get('articles'), channel_id=r.get('channel_id'))
-        delete_articles(channel_id=r.get("channel_id"))
-        articles.extend(ar)
-    Article.objects.bulk_create(articles)
+    updating = is_user_channels_updating(user_id=user_id)
+    print(updating)
+    if updating:
+        return
+    set_user_channels_updating(user_id=user_id, insert=True)
+    try:
+        user = ApplicationUser.objects.get(id=user_id)
+        channels = get_user_existing_channels(user=user)
+        g = group(fetch_articles_task.s(c.rss_link, c.id) for c in channels)
+        result = g().get()
+        Article.objects.filter(channel__in=channels).delete()
+        articles = []
+        for r in result:
+            if r is None:
+                continue
+            ar = _create_articles_from_task_result(result_list=r.get('articles'), channel_id=r.get('channel_id'))
+            delete_articles(channel_id=r.get("channel_id"))
+            articles.extend(ar)
+        Article.objects.bulk_create(articles)
+    finally:
+        set_user_channels_updating(user_id=user_id, insert=False)
 
 
 @shared_task
 def update_channel(user_id: int, chanel_id: int):
-    user = get_user_by_id(user_id=user_id)
-    channel = get_channel_by_id(user=user, subscription_id=chanel_id)
-    result = fetch_articles_task.delay(channel.rss_link, chanel_id).get()
-    if result is None:
+    updating = is_user_channel_updating(user_id=user_id, channel_id=chanel_id)
+    print(updating)
+    if updating:
         return
-    articles = _create_articles_from_task_result(result.get('articles'), channel_id=channel.id)
-    if articles is not None and len(articles) == 0:
-        delete_articles(channel_id=channel.id)
-        Article.objects.bulk_create(articles)
+    set_user_channel_updating(user_id=user_id, channel_id=chanel_id, insert=True)
+    try:
+        user = get_user_by_id(user_id=user_id)
+        channel = get_channel_by_id(user=user, subscription_id=chanel_id)
+        result = fetch_articles_task.delay(channel.rss_link, chanel_id).get()
+        if result is None:
+            return
+        articles = _create_articles_from_task_result(result.get('articles'), channel_id=channel.id)
+        if articles is not None and len(articles) == 0:
+            delete_articles(channel_id=channel.id)
+            Article.objects.bulk_create(articles)
+    finally:
+        set_user_channel_updating(user_id=user_id, channel_id=chanel_id, insert=False)
 
 
 def _create_articles_from_task_result(result_list: List[dict], channel_id: int, **kwargs) -> Optional[List[Article]]:
